Initial commit of ABFRealmSearchViewController and example project.
ABFRealmSearchViewController is a subclass of UITableViewController. Similarly, it manages the display of a table view to dispaly search results and update based off of text input into the search bar. The controller can be customized through various properties, including leaving the search bar out of the tableview header and inserting it elsewhere as appropriate.